requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
streamlit-drawable-canvas>=0.9.3
pillow>=10.0.0
